+++ front/tests/utils/AgentConfigurationFactory.ts
@@ -68,37 +68,4 @@ export class AgentConfigurationFactory {
 
     return result.value;
   }
-
-  static async createTestAgentWithMCPAction(
-    auth: Authenticator,
-    t: Transaction
-  ): Promise<LightAgentConfigurationType> {
-    const owner = auth.getNonNullableWorkspace();
-
-    const agent = await AgentConfigurationFactory.createTestAgent(auth, t);
-    const mcpServerView = await MCPServerViewFactory.create(
-      owner,
-      "dummy_mcp_server_id",
-      await SpaceFactory.global(owner, t)
-    );
-
-    await AgentMCPServerConfiguration.create(
-      {
-        sId: generateRandomModelSId(),
-        agentConfigurationId: agent.id,
-        workspaceId: owner.id,
-        mcpServerViewId: mcpServerView.id,
-        internalMCPServerId: "internal_mcp_server_id",
-        additionalConfiguration: {},
-        timeFrame: null,
-        jsonSchema: null,
-        name: null,
-        singleToolDescriptionOverride: null,
-        appId: null,
-      },
-      { transaction: t }
-    );
-
-    return agent;
-  }
 }
